fix for "add 2 items one after each other" issue
