+++ src/selectors/backlogItemSelectors.ts
@@ -4,8 +4,10 @@ import type { BacklogItemsState } from "../reducers/backlogItems/backlogItemsRed
 import type { StateTree } from "../reducers/rootReducer";
 
 // utils
-import { getBacklogItemById as reducerGetBacklogItemById } from "../reducers/backlogItems/backlogItemsReducerHelper";
 import { createSelector } from "reselect";
+
+// selectors
+import * as backlogItemsSliceSelectors from "../reducers/backlogItems/backlogItemsSliceSelectors";
 
 export const getBacklogItemByInstanceId = (state: StateTree, instanceId: number): BacklogItem | null => {
     const matchingItems = state.backlogItems.addedItems.filter((addedItem) => addedItem.instanceId === instanceId);
@@ -17,9 +19,8 @@ export const getBacklogItemByInstanceId = (state: StateTree, instanceId: number)
     }
 };
 
-export const getBacklogItemById = (state: StateTree, itemId: string): BacklogItem | null => {
-    return reducerGetBacklogItemById(state.backlogItems, itemId);
-};
+export const getBacklogItemById = (state: StateTree, itemId: string): BacklogItem | null =>
+    backlogItemsSliceSelectors.getBacklogItemById(state.backlogItems, itemId);
 
 export const getCurrentBacklogItem = (state: StateTree): BacklogItem | null => {
     return state.backlogItems.currentItem;
@@ -84,9 +85,8 @@ export const getPrevNextAndCurrentById = (state: StateTree, id: string): PrevNex
 
 export const getAllBacklogItems = (state: StateTree) => state.backlogItems.allItems;
 
-export const getSelectedBacklogItemIds = (state: StateTree): string[] => {
-    return state.backlogItems.selectedItemIds;
-};
+export const getSelectedBacklogItemIds = (state: StateTree): string[] =>
+    backlogItemsSliceSelectors.getSelectedBacklogItemIdsFromSlice(state.backlogItems);
 
 export const getSelectedBacklogItems = (state: StateTree) => {
     const results = [];
